requirements.txt
torch>=2.0.0
transformers>=4.28.0
Pillow>=9.0.0
pydantic>=2.0.0
